requirements.txt
pip
aiohttp
discord.py
pytest
selenium